requirements.txt
streamlit
openai
python-dotenv
psycopg2-binary
supabase
requests